consumer/requirements.txt
confluent_kafka
supabase
python-dotenv
msgpack